requirements.txt
asgiref==3.4.1
certifi==2022.5.18.1
charset-normalizer==2.0.12
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==1.0.0
Django==4.0.1
django-cors-headers==3.13.0
django-filter==21.1
django-heroku==0.3.1
djangorestframework==3.13.1
djangorestframework-simplejwt==5.2.0
drf-yasg==1.20.0
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.10.1
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
Markdown==3.3.6
MarkupSafe==2.1.1
packaging==21.3
Pillow==9.1.1
psycopg2==2.9.5
PyJWT==2.4.0
pyparsing==3.0.9
pytz==2021.3
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
sqlparse==0.4.2
tzdata==2021.5
uritemplate==4.1.1
urllib3==1.26.9
whitenoise==6.2.0
zipp==3.7.0
